Release v1.0.5

+++ app/gen-template.sh
@@ -1,21 +1,22 @@
 #!/bin/bash
-##
-## Usage: ./gen-template.sh [username ...]
-##
-##    e.g ./gen-template.sh $(cat member.list)
-##     or ./gen-template.sh user01 user02
-##
+
+## Usage: ./gen-template.sh $(cat member.list)
 
 BASEDIR=$(dirname $(readlink -f $0))
 OUTPUT_DIRNAME="$(date +%Y%m%d.%H%M%S)"
 OUTPUTDIR="${DATADIR:-/root}/${OUTPUT_DIRNAME}"
 
-OPENSSL_CMD="openssl"
-MC_CMD=${BASEDIR}/mc
-MINIO_BUCKET=${MINIO_BUCKET:-my-minio-local}
+MC_CMD="${BASEDIR}/mc"
+MINIO_TARGET=${MINIO_TARGET:-my-minio-local}
 
 ## template filepath from ${OUTPUTDIR}
 TEMPLPATH1="${BASEDIR}/user-bucket-policy-_USER_ID_.json"
+
+## check the .mc/config file.
+if test ! -f ~/.mc/config.json || test $(grep "${MINIO_TARGET}" ~/.mc/config.json >/dev/null) == "1" ; then
+  echo "Usage: Please setup the ~/.mc/config file, using the following command line."
+  echo "  ./mc config host add my-minio-local http://my-minio.minio:9000 <root-access-key> <root-password> --api S3v4"
+fi
 
 ## prepare outputdir and cd to it.
 mkdir -p "${OUTPUTDIR}"
@@ -23,15 +24,21 @@ cd "${OUTPUTDIR}"
 
 for username in $@
 do
+    ## check existing bucket
+    ${MC_CMD} ls "${MINIO_TARGET}/${username}-bucket"
+    test "$?" == "0" && echo "[skip] bucket, ${username}-bucket, exists" && continue
+
+    ## prepare the template file
     OUTPUT_FILENAME=$(basename "${TEMPLPATH1}" | sed -e "s/_USER_ID_/${username}/g")
     sed -e "s/_USER_ID_/${username}/g" "${TEMPLPATH1}" > "${OUTPUT_FILENAME}"
 
-    ACCESS_KEY="$(${OPENSSL_CMD} rand -hex 8)"
-    SECRET_KEY="$(${OPENSSL_CMD} rand -hex 16)"
-    ${MC_CMD} admin policy add "${MINIO_BUCKET}" "readwrite-${username}" "${OUTPUT_FILENAME}"
-    test "$?" == "0" && ${MC_CMD} admin user add "${MINIO_BUCKET}" "${ACCESS_KEY}" "${SECRET_KEY}"
-    test "$?" == "0" && ${MC_CMD} admin policy set "${MINIO_BUCKET}" "readwrite-${username}" user="${ACCESS_KEY}"
+    ## adding minio user and policy
+    ACCESS_KEY="$(openssl rand -hex 8)"
+    SECRET_KEY="$(openssl rand -hex 16)"
+    ${MC_CMD} admin policy create "${MINIO_TARGET}" "readwrite-${username}" "${OUTPUT_FILENAME}"
+    test "$?" == "0" && ${MC_CMD} admin user add "${MINIO_TARGET}" "${ACCESS_KEY}" "${SECRET_KEY}"
+    test "$?" == "0" && ${MC_CMD} admin policy attach "${MINIO_TARGET}" "readwrite-${username}" --user "${ACCESS_KEY}"
     echo "${username},${ACCESS_KEY},${SECRET_KEY}" >> "../${OUTPUT_DIRNAME}.txt"
 
-    ${MC_CMD} mb "${MINIO_BUCKET}/${username}-bucket"
+    ${MC_CMD} mb "${MINIO_TARGET}/${username}-bucket"
 done
